requirements.txt
Pillow
webp